@@ -1,23 +1,22 @@
 name: aws_client
 description: High-level APIs for Amazon Web Services (AWS)
-version: 0.3.0-dev
-author: Istvan Soos <istvan.soos@agilord.com>
-
+version: 1.0.0
+# author: Istvan Soos <istvan.soos@agilord.com>
 homepage: https://github.com/agilord/aws_client
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
-  crypto: ^2.1.3
-  http_client: '>=0.5.0 <2.0.0'
-  meta: ^1.1.8
-  xml: ^3.5.0
-  json_annotation: ^3.0.0
+  crypto: ^3.0.2
+  http_client: ^1.5.1
+  json_annotation: ^4.8.0
+  meta: ^1.9.0
+  xml: ^6.2.2
 
 dev_dependencies:
-  pedantic: ^1.8.0+1
-  test: ^1.9.4
-  archive: ^2.0.11
-  build_runner: ^1.7.2
-  json_serializable: ^3.2.3
+  archive: ^3.3.6
+  build_runner: ^2.3.3
+  json_serializable: ^6.6.0
+  lints: ^2.0.1
+  test: ^1.22.2
