@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file.
 
-import 'package:meta/meta.dart';
-
 /// AWS credentials.
 class Credentials {
   /// AWS access key
@@ -13,12 +11,11 @@ class Credentials {
   final String secretKey;
 
   /// AWS temporary credentials session token
-  final String sessionToken;
+  final String? sessionToken;
 
   /// AWS credentials.
   Credentials(
-      {@required this.accessKey, @required this.secretKey, this.sessionToken}) {
-    assert(accessKey != null);
-    assert(secretKey != null);
-  }
+      {required this.accessKey, required this.secretKey, this.sessionToken})
+      : assert(accessKey != '', 'accessKey must not be empty'),
+        assert(secretKey != '', 'secretKey must not be empty');
 }
